Switch to using Module.prepend instead of method aliasing in redis hook
Update specs to test if the prepend works by removing the stub from the
`process` call (previously `process_without_appsignal`).

If we don't change the implementation to `super` instead of
`process_without_appsignal`, the spec still passed.
